% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addMetaData}
\alias{addMetaData}
\title{Add meta information of CYT}
\usage{
addMetaData(object, meta.info, name = "NewCol", verbose = FALSE)
}
\arguments{
\item{object}{A CYT object}

\item{meta.info}{a vector, meta data of cell information}

\item{name}{character, colname of `meta.info`}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object
}
\description{
Add meta information of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

plot.meta <- fetchPlotMeta(cyt)
meta.info <- 1:nrow(plot.meta)
names(meta.info) <- plot.meta$cell

cyt <- addMetaData(cyt, meta.info = meta.info, name = "MyInformation")
plot.meta <- fetchPlotMeta(cyt)


}
