% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{buildTree}
\alias{buildTree}
\title{buildTree}
\usage{
buildTree(
  object,
  method = "euclidean",
  dim.type = c("raw", "pca", "tsne", "dc", "umap"),
  dim.use = seq_len(2),
  verbose = FALSE
)
}
\arguments{
\item{object}{a CYT object}

\item{method}{character. Mehtod to build MST.}

\item{dim.type}{character. Type of dimensions that will be used to build the tree.
Five \code{dim.type} are provided, 'raw', 'pca', 'tsne', 'dc' and 'umap'. 
By default is 'raw'.}

\item{dim.use}{numeric. Number of dimensions that will be used to build the tree.
For example. If \code{dim.use} is 'raw', there is no limit for \code{dim.type}. 
And if the \code{dim.use} is 'tsne' or 'umap', the default \code{dim.use} is seq_len(2).}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object with tree
}
\description{
buildTree
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

cyt <- buildTree(cyt, dim.type = "raw")

# build minimum spanning tree (MST) based on tsne
cyt <- buildTree(cyt, dim.type = "tsne", dim.use = seq_len(2))

# Using PCA
cyt <- buildTree(cyt, dim.type = "pca", dim.use =seq_len(4))

# Using UMAP
cyt <- buildTree(cyt, dim.type = "umap", dim.use = seq_len(2))

# Using Diffusion Maps
cyt <- buildTree(cyt, dim.type = "dc", dim.use = seq_len(3))


}
