% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeMarker}
\alias{changeMarker}
\title{Change marker used in the calculation of CYT}
\usage{
changeMarker(object, markers = NULL, verbose = FALSE)
}
\arguments{
\item{object}{A CYT object}

\item{markers}{vector. Markers used in the calculation
of PCA, tSNE, diffusion map and UMAP.}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object
}
\description{
Change marker used in the calculation of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

markers <- c("CD43", "CD34", "CD90", "CD45RA")

cyt <- changeMarker(cyt, markers = markers)


      
}
