% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constraintMatrix}
\alias{constraintMatrix}
\title{constraintMatrix}
\usage{
constraintMatrix(x, cutoff = 0.99, markers = NULL, method = "euclidean")
}
\arguments{
\item{x}{matrix}

\item{cutoff}{numeric. Cutoff of the constraint value}

\item{markers}{character. Markers used in the calculation of constraint model.}

\item{method}{character. the distance measure to be used.
This must be one of "euclidean", "maximum", "manhattan",
"canberra", "binary" or "minkowski".}
}
\value{
a matrix
}
\description{
constraint FCS data by a provid cutoff
}
\examples{

mat <- matrix(runif(10000), nrow = 1000, ncol = 10)
colnames(mat) <- LETTERS[1:10]
dim(mat)

mat <- constraintMatrix(mat)
dim(mat)

}
