% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{correctBatchCYT}
\alias{correctBatchCYT}
\title{correctBatchCYT}
\usage{
correctBatchCYT(
  object,
  batch = NULL,
  par.prior = TRUE,
  mean.only = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An CYT object}

\item{batch}{vector. Batch covariate (only one batch allowed)}

\item{par.prior}{logical. TRUE indicates parametric adjustments will be used,
FALSE indicates non-parametric adjustments will be used.}

\item{mean.only}{logical. FALSE If TRUE ComBat only corrects the mean of the batch
effect (no scale adjustment)}

\item{verbose}{logical. Whether to show log information}

\item{...}{Parameters passing to \code{\link[sva]{ComBat}} function}
}
\value{
An CYT object after removing batch effect

An CYT object with corrected batch effects
}
\description{
Remove batch effect in CYT object
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)
plot.meta <- fetchPlotMeta(cyt)
batch <- as.numeric(plot.meta$stage)
cyt <- correctBatchCYT(cyt, batch = batch)


}
\seealso{
\code{\link[BiocNeighbors]{findKNN}}
}
