% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{correctBatchFSPY}
\alias{correctBatchFSPY}
\title{correctBatchFSPY}
\usage{
correctBatchFSPY(
  object,
  batch = NULL,
  par.prior = TRUE,
  mean.only = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An FSPY object}

\item{batch}{vector. Batch covariate (only one batch allowed)}

\item{par.prior}{logical. TRUE indicates parametric adjustments will be used,
FALSE indicates non-parametric adjustments will be used.}

\item{mean.only}{logical. FALSE If TRUE ComBat only corrects the mean of the batch
effect (no scale adjustment)}

\item{verbose}{logical. Whether to show log information}

\item{...}{Parameters passing to \code{\link[sva]{ComBat}} function}
}
\value{
An FSPY object after removing batch effect

An FSPY object with corrected batch effects
}
\description{
Remove batch effect in FSPY object
}
\examples{

if (FALSE) {
  plot.meta <- fetchPlotMeta(fspy)
  batch <- as.numeric(plot.meta$stage)
  fspy <- correctBatchFSPY(object, batch = batch)
}

}
\seealso{
\code{\link[BiocNeighbors]{findKNN}}
}
