% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defLeafCells}
\alias{defLeafCells}
\title{definition of leaf cells}
\usage{
defLeafCells(object, leaf.cells = NULL, pseudotime.cutoff = 0, verbose = FALSE)
}
\arguments{
\item{object}{a CYT object}

\item{leaf.cells}{character or numeric. Cell name of the root cells or
cluster.id of root.cells}

\item{pseudotime.cutoff}{numeric. Cutoff of pseudotime. Cells with pseudotime
over pseudotime.cutoff will be set to be leaf cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object
}
\description{
definition of root cells
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

# Define leaf cells by cluster
cyt <- defLeafCells(cyt, leaf.cells = 1, verbose = TRUE)
cyt <- defLeafCells(cyt, leaf.cells = c(1,3), verbose = TRUE)

# Define root cells by cell names
meta.data <- fetchPlotMeta(cyt)
cells <- meta.data$cell[which(meta.data$stage == "D10")]
cells <- as.character(cells)
cyt <- defLeafCells(cyt, leaf.cells = cells, verbose = TRUE)



}
