% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchClustMeta}
\alias{fetchClustMeta}
\title{Fetching clusters' metadata of CYT}
\usage{
fetchClustMeta(object, verbose = FALSE)
}
\arguments{
\item{object}{An CYT object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
a data.frame containing clustering information for visualization
}
\description{
Fetching clusters' metadata of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

clust.data <- fetchClustMeta(cyt)
head(clust.data)



}
