% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gatingMatrix}
\alias{gatingMatrix}
\title{Apply gating on the matrix data}
\usage{
gatingMatrix(x, lower.gate = NULL, upper.gate = NULL)
}
\arguments{
\item{x}{matrix}

\item{lower.gate}{vector. Gating parameter, the name of the vector is the marker name, and
the value of the vector is the lower bound of gating cutoff.}

\item{upper.gate}{vector. Gating parameter, the name of the vector is the marker name, and
the value of the vector is the upper bound of gating cutoff.}
}
\value{
a matrix
}
\description{
Apply gating on the matrix data
}
\examples{
par(mfrow=c(1,2))
x <- matrix(rnorm(200, 3, 1), nrow = 100, ncol = 2)
colnames(x) <- c("CD34", "CD43")
plot(x[, "CD34"], x[, "CD43"], main = "Before gating")

lower.gate = c(CD34 = 2, CD43 = 3)
upper.gate = c(CD34 = 4, CD43 = 5)

x <- gatingMatrix(x, lower.gate = lower.gate, upper.gate = upper.gate)
plot(x[, "CD34"], x[, "CD43"], main = "After gating")

par(mfrow=c(1,1))


}
