% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotBranchHeatmap}
\alias{plotBranchHeatmap}
\title{Visualization heatmap of branch data of CYT}
\usage{
plotBranchHeatmap(
  object,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  ...
)
}
\arguments{
\item{object}{An CYT object}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of branch data of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

plotBranchHeatmap(cyt)
plotBranchHeatmap(cyt, color = colorRampPalette(c("purple","white","yellow"))(100))
plotBranchHeatmap(cyt, cluster_row = FALSE)
plotBranchHeatmap(cyt, cluster_row = FALSE, cluster_col = FALSE)



}
