% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotCluster}
\alias{plotCluster}
\title{Visualization of cluster data of CYT}
\usage{
plotCluster(
  object,
  item.use = c("PC_1", "PC_2"),
  color.by = "cluster",
  size.by = "cell.number.percent",
  order.by = NULL,
  size = 1,
  alpha = 1,
  category = "categorical",
  show.cluser.id = FALSE,
  show.cluser.id.size = 4,
  main = "2D plot of cluster in CYT",
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{An CYT object}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}

\item{size.by}{character. Size of the dot}

\item{order.by}{vector. Order of color theme.}

\item{size}{numeric. Size of the dot}

\item{alpha}{numberic. Transparency (0-1) of the dot, default is 1.}

\item{category}{character. numeric or categorical}

\item{show.cluser.id}{logical. Whether to show cluster id in the plot.}

\item{show.cluser.id.size}{numeric. Size of the cluster id.}

\item{main}{character. Title of the plot.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
Visualization of cluster data of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

plotCluster(cyt)

plotCluster(cyt, item.use = c("PC_1", "PC_2"))
plotCluster(cyt, item.use = c("PC_2", "PC_3"))
plotCluster(cyt, item.use = c("PC_2", "PC_3"), color.by = "CD43", category = "numeric")
plotCluster(cyt, item.use = c("PC_2", "PC_3"), color.by = "CD43", category = "numeric")

plotCluster(cyt, item.use = c("tSNE_1", "tSNE_2"))
plotCluster(cyt, item.use = c("tSNE_1", "tSNE_2"), show.cluser.id = TRUE)

plotCluster(cyt, item.use = c("DC_1", "DC_2"))

plotCluster(cyt, item.use = c("UMAP_1", "UMAP_2"))


}
