% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Visualization heatmap of cluster data of CYT}
\usage{
plotClusterHeatmap(
  object,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  ...
)
}
\arguments{
\item{object}{An CYT object}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of cluster data of CYT
}
\examples{

if (FALSE) {

plotClusterHeatmap(cyt)
plotClusterHeatmap(cyt, color = colorRampPalette(c("purple","white","yellow"))(100))
plotClusterHeatmap(cyt, cluster_row = F)
plotClusterHeatmap(cyt, cluster_row = F, cluster_col = F)

}

}
