% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotPieTree}
\alias{plotPieTree}
\title{plot MST pie of FSPY}
\usage{
plotPieTree(
  object,
  cex.size = 2,
  size.by.cell.number = TRUE,
  as.tree = FALSE,
  root.id = NULL,
  show.node.name = FALSE
)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{size.by.cell.number}{logical. Whether to size node by cell number}

\item{as.tree}{logical. Whether to show node as tree}

\item{root.id}{numeric. Root id of the tree, if as.tree is TRUE}

\item{show.node.name}{logical. whether to show node name}
}
\value{
ggplot2 figure
}
\description{
plot MST pie of FSPY
}
\examples{

if (FALSE) {

# Runs only have two or more stages
plotPieTree(fspy, cex.size = 1, size.by.cell.number = T) +
   scale_fill_manual(values = c("#00599F","#FF3222","#009900",
                                "#FF9933","#FF99FF","#7A06A0"))
}

}
