% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plotPseudotimeDensity}
\alias{plotPseudotimeDensity}
\title{plot Pseudotime density of CYT}
\usage{
plotPseudotimeDensity(
  object,
  color.by = "stage",
  main = "Density of pseudotime",
  adjust = 0.5,
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{an CYT object}

\item{color.by}{character.}

\item{main}{character. Title of the plot}

\item{adjust}{numeric. A multiplicate bandwidth adjustment.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
plot Pseudotime density of CYT
}
\examples{

if (FALSE) {

plotPseudotimeDensity(cyt)

plotPseudotimeDensity(cyt, adjust = 1)
plotPseudotimeDensity(cyt, adjust = 2)

plotPseudotimeDensity(cyt, adjust = 2) +
  scale_color_manual(values = c("#00599F","#FF3222","#009900",
                                "#FF9933","#FF99FF","#7A06A0"))

}

}
