% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCYT.R
\name{plotPseudotimeTraj}
\alias{plotPseudotimeTraj}
\title{plotPseudotimeTraj}
\usage{
plotPseudotimeTraj(
  object,
  cutoff = -1,
  markers = NULL,
  size = 0.5,
  alpha = 0.6,
  print.curve = TRUE,
  var.cols = FALSE,
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{An CYT object}

\item{cutoff}{numeric. Cutoff of trajectory value}

\item{markers}{character. Markers used in the calculation progress}

\item{size}{numeric. Size of the dot}

\item{alpha}{numeric. Transparency (0-1) of the dot, default is 1.}

\item{print.curve}{logical. Whether to perform curve fitting}

\item{var.cols}{logical. Whether to plot stage}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
plotPseudotimeTraj
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

plotPseudotimeTraj(cyt)
plotPseudotimeTraj(cyt, print.curve = FALSE)
plotPseudotimeTraj(cyt, var.cols = TRUE)

plotPseudotimeTraj(cyt, markers = c("CD43", "CD34"))



}
