% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotTrajHeatmap}
\alias{plotTrajHeatmap}
\title{Visualization heatmap of intermediate cells of CYT}
\usage{
plotTrajHeatmap(
  object,
  cutoff = 0,
  markers = NULL,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  ...
)
}
\arguments{
\item{object}{A CYT object}

\item{cutoff}{numeric. value to identify intermediate state cells}

\item{markers}{markers to plot on the heatmap}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of intermediate cells of CYT
}
