% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plot MST of FSPY}
\usage{
plotTree(
  object,
  cex.size = 1,
  color.by = "cell.number",
  size.by = "cell.number",
  as.tree = FALSE,
  root.id = NULL,
  show.node.name = FALSE
)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{color.by}{numeric. size color theme of the dot}

\item{size.by}{numeric. size theme of the dot}

\item{as.tree}{logical. Whether to show node as tree}

\item{root.id}{numeric. Root id of the tree, if as.tree is TRUE}

\item{show.node.name}{logical. whether to show node name}
}
\value{
ggplot2 figure
}
\description{
plot MST of FSPY
}
\examples{

if (FALSE) {

plotTree(fspy)

plotTree(fspy, show.node.name = T)

plotTree(fspy, color.by = "CD43", show.node.name = T, cex.size = 1) +
    scale_colour_gradientn(colors = c("#00599F", "#EEEEEE", "#FF3222"))

plotTree(fspy, color.by = "D0.percent", show.node.name = T, cex.size = 1) +
    scale_colour_gradientn(colors = c("#00599F", "#EEEEEE", "#FF3222"))

plotTree(fspy, color.by = "D2.percent", show.node.name = T, cex.size = 1) +
    scale_colour_gradientn(colors = c("#00599F", "#EEEEEE", "#FF3222"))

plotTree(fspy, color.by = "pseudotime", cex.size = 1) +
    scale_colour_gradientn(colors = c("#F4D31D", "#FF3222","#7A06A0"))

}

}
