% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotViolin}
\alias{plotViolin}
\title{Visualization violin plot of CYT}
\usage{
plotViolin(
  object,
  marker,
  color.by = "cluster.id",
  order.by = NULL,
  size = 1,
  text.angle = 0,
  main = "Violin plot CYT",
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{A CYT object}

\item{marker}{character. Markers used to plot}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}

\item{order.by}{vector. Order of color theme.}

\item{size}{numeric. Size of the dot}

\item{text.angle}{numberic. Text angle of the violin plot}

\item{main}{character. Title of the plot.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
Visualization violin plot of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

plotViolin(cyt, marker = "CD34")
plotViolin(cyt, marker = "CD34", order.by = "pseudotime")


}
