% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{processingCluster}
\alias{processingCluster}
\title{processingCluster}
\usage{
processingCluster(
  object,
  perplexity = 5,
  k = 5,
  downsampling.size = 1,
  force.resample = TRUE,
  random.cluster = FALSE,
  umap.config = umap.defaults,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{perplexity}{numeric. Perplexity parameter (should not be bigger than 3 *
perplexity < nrow(X) - 1, see details for interpretation). See \code{\link[Rtsne]{Rtsne}}
for more information.}

\item{k}{numeric. The parameter k in k-Nearest Neighbor.}

\item{downsampling.size}{numeric. Percentage of sample size of downsampling.
This parameter is from 0 to 1. by default is 1.}

\item{force.resample}{logical. Whether to do resample if downsampling.size < 1}

\item{random.cluster}{logical. Whether to perfrom random downsampling. If FALSE, 
an uniform downsampling will be processed.}

\item{umap.config}{object of class umap.config. See \code{\link[umap]{umap}}.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the dimensionality reduction functions.}
}
\value{
An FSPY object with cluster.id in meta.data

An FSPY object with dimensionality reduction of clusters
}
\description{
Calculate Principal Components Analysis (PCA), t-Distributed
   Stochastic Neighbor Embedding (tSNE), Diffusion Map and Uniform Manifold
   Approximation and Projection (UMAP) of clusters calculated by runCluster.
}
\examples{

if (FALSE) {

# After running clustering
set.seed(1)
fspy <- runCluster(fspy, cluster.method = "som", xdim = 3, ydim = 3, verbose = T)

# Do not perfrom downsampling
fspy <- processingCluster(fspy, perplexity = 2)

# Perform cluster based downsampling
# Only keep 50\% cells
fspy <- processingCluster(fspy, perplexity = 2, downsampling.size = 0.5)

# Processing clusters without downsampling step
fspy <- processingCluster(fspy, perplexity = 2, force.resample = FALSE)

}

}
\seealso{
\code{\link[umap]{umap}}, \code{\link[gmodels]{fast.prcomp}},
   \code{\link[Rtsne]{Rtsne}}, \code{destiny}
}
