% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runClara}
\alias{runClara}
\title{runClara}
\usage{
runClara(
  object,
  k = 25,
  metric = c("euclidean", "manhattan", "jaccard"),
  stand = FALSE,
  samples = 5,
  scale = TRUE,
  trace = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{k}{numeric. The number of clusters. It is required that
0 < k < n where n is the number of observations (i.e., n = nrow(x)).}

\item{metric}{character. string specifying the metric to be used for
calculating dissimilarities between observations.}

\item{stand}{logical. Indicating if the measurements in x are
standardized before calculating the dissimilarities.}

\item{samples}{numeric. Say N, the number of samples to be drawn from the dataset.
The default is N = 5,}

\item{scale}{logical. Whether to use scaled data in kmeans.}

\item{trace}{numberic. Indicating a trace level for diagnostic output during the algorithm}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[cluster]{clara}} function}
}
\value{
an FSPY object with clara.id in meta.data
}
\description{
Clustering a data matrix into k clusters
}
\examples{
if (FALSE) {
fspy <- runClara(fspy, k = 25, verbose = TRUE)
}

}
\seealso{
\code{\link[cluster]{clara}}
}
