% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runCluster}
\alias{runCluster}
\title{Specific Clustering Method Toolkits}
\usage{
runCluster(
  object,
  cluster.method = c("som", "kmeans", "clara", "phenograph", "hclust", "mclust"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{a CYT object}

\item{cluster.method}{character. Four clustering method are provided: som, clara, kmeans and phenograph.
Clustering method "hclust" and "mclust" are not recommended because of long computing time.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
A CYT object with cluster
}
\description{
Compute a specific clustering using the combined flow
   cytometry data. "som" \code{\link[FlowSOM]{SOM}}, "hclust" \code{\link[stats]{hclust}},
   "clara" \code{\link[cluster]{clara}}, "phenograph", "kmeans" \code{\link[stats]{kmeans}} are
   provided.
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

# After building an CYT object
# Set random seed to make results reproducible

set.seed(1)
cyt <- runCluster(cyt, cluster.method = "som", xdim = 3, ydim = 3, verbose = TRUE)

# K-means clustering
cyt <- runCluster(cyt, cluster.method = "kmeans", k = 9, verbose = TRUE)

# Clara clustering
cyt <- runCluster(cyt, cluster.method = "clara", k = 9, verbose = TRUE)

# phenoGraph clustering
cyt <- runCluster(cyt, cluster.method = "phenograph", verbose = TRUE)

# hclust clustering
# not recommended for large cell size
cyt <- runCluster(cyt, cluster.method = "hclust", k = 9, verbose = TRUE)

# mclust clustering
# not recommended for large cell size
cyt <- runCluster(cyt, cluster.method = "mclust", verbose = TRUE)


}
\seealso{
\code{\link[FlowSOM]{SOM}}, \code{\link[stats]{hclust}},
   \code{\link[cluster]{clara}}, \code{\link[stats]{kmeans}}.
   You can use \code{runSOM}, \code{runClara},
   \code{runPhenotype}, \code{runKmeans}, \code{runMclust} and
   \code{runHclust} to run clustering respectively.
}
