% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiff.R
\name{runDiff}
\alias{runDiff}
\title{Calculate differential expression markers}
\usage{
runDiff(object, branch.id = NULL, branch.id.2 = NULL, verbose = FALSE)
}
\arguments{
\item{object}{an FSPY object}

\item{branch.id}{vector. Branch ids use to run differentially expressed markers}

\item{branch.id.2}{vector. Branch ids use to run differentially expressed
markers in compare with branch.id}

\item{verbose}{logic. Whether to print calculation progress.}
}
\value{
An FSPY object with cluster.id in meta.data

a data.frame with differential expressed markers
}
\description{
Calculating differentially expressed markers
}
\examples{

if (FALSE) {

DEG.table <- runDiff(fspy)

}



}
\seealso{
\code{bulidTree}
}
