% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destiny.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\title{Calculate diffusion map in FSPY}
\usage{
runDiffusionMap(
  object,
  sigma.use = NULL,
  distance = c("euclidean", "cosine", "rankcor"),
  k = 30,
  density.norm = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{sigma.use}{numeric. Diffusion scale parameter of the Gaussian kernel. One of '\code{local}',
 '\code{global}', a \code{\link[base]{numeric}} global sigma or a Sigmas object.
 When choosing '\code{global}', a global sigma will be calculated using find_sigmas
 (See \code{destiny}). A larger sigma might be necessary if the eigenvalues can not
be found because of a singularity in the matrix. See \code{destiny}.}

\item{distance}{Distance measurement method applied to data or a distance matrix/dist.
For the allowed values, see \code{destiny}}

\item{k}{numeric. By default is 30. \code{destiny} can be used to specify k.}

\item{density.norm}{logical. If TRUE, use density normalisation. See \code{destiny}}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{options to pass on to the \code{destiny}.}
}
\value{
An FSPY object
}
\description{
Calculate diffusion map in FSPY
}
\examples{

if (FALSE) {
fspy <- runDiffusionMap(fspy, verbose = TRUE)
}

}
\seealso{
\code{destiny}
}
