% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{runExprsExtract}
\alias{runExprsExtract}
\title{Extract the expression data from a FCS file with preprocessing}
\usage{
runExprsExtract(
  fcsFile,
  verbose = FALSE,
  comp = FALSE,
  transformMethod = c("autoLgcl", "cytofAsinh", "logicle", "arcsinh", "logAbs", "none"),
  scaleTo = NULL,
  showDesc = TRUE,
  keepRaw = TRUE,
  q = 0.05,
  l_w = 0.1,
  l_t = 4000,
  l_m = 4.5,
  l_a = 0,
  a_a = 1,
  a_b = 1,
  a_c = 0
)
}
\arguments{
\item{fcsFile}{The name of the FCS file.}

\item{verbose}{If \verb{TRUE}, print the message details of FCS loading.}

\item{comp}{If \verb{TRUE}, does compensation  by compensation matrix
contained in FCS. Agrument also accepts a compensation matrix to be
applied. Otherwise \verb{FALSE}.}

\item{transformMethod}{Data Transformation method, including \code{autoLgcl},
\code{cytofAsinh}, \code{logicle} and \code{arcsinh}, or \code{none}
to avoid transformation.}

\item{scaleTo}{Scale the expression to a specified range c(a, b), default is NULL.}

\item{showDesc}{logical. Whether to show \code{desc} name in the output matrix.}

\item{keepRaw}{logical. Whether to keep raw data for FSC and SSC.}

\item{q}{Quantile of negative values removed for auto w estimation,
default is 0.05, parameter for autoLgcl transformation.}

\item{l_w}{Linearization width in asymptotic decades, parameter for
logicle transformation.}

\item{l_t}{Top of the scale data value, parameter for logicle transformation.}

\item{l_m}{Full width of the transformed display in asymptotic decades,
parameter for logicle transformation.}

\item{l_a}{Additional negative range to be included in the display
in asymptotic decades, parameter for logicle transformation.}

\item{a_a}{Positive double that corresponds to the base of the arcsinh
transformation, \code{arcsinh} = asinh(a + b * x) + c).}

\item{a_b}{Positive double that corresponds to a scale factor of the
arcsinh transformation, \code{arcsinh} = asinh(a + b * x) + c).}

\item{a_c}{Positive double that corresponds to another scale factor
of the arcsinh transformation, \code{arcsinh} = asinh(a + b * x) + c).}
}
\value{
A transformed expression data matrix
}
\description{
Extract the FCS expresssion data with preprocessing of
   compensation (for FCM data only) and transformation. Transformtion
   methods includes \code{autoLgcl}, \code{cytofAsinh}, \code{logicle}
   (customizable) and \code{arcsinh} (customizable).
}
\examples{

if (FALSE) {
# See vignette tutorials for more information
vignette(package = "flowSpy")
vignette("Quick_start", package = "flowSpy")

  # Path to your FCS files
  fcs.path <- "flowSpy-dataset/FCS/usecase1/"
  fcs.file <- paste0(fcs.path, "FR-FCM-ZY9R-Bone_Marrow_cytof.fcs")

  # Read FCS files
  exp.data <- runExprsExtract(fcs.file, showDesc = FALSE, transformMethod = "autoLgcl")
}


}
\references{
Hao Chen, Mai Chan Lau, Michael Thomas Wong, Evan W. Newell,
   Michael Poidinger, Jinmiao Chen. Cytofkit: A Bioconductor Package for
   an Integrated Mass Cytometry Data Analysis Pipeline. PLoS Comput Biol, 2016.
}
\author{
Chen Hao
}
