% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{runExprsMerge}
\alias{runExprsMerge}
\title{Merge the expression matrix from multiple FCS files with preprocessing}
\usage{
runExprsMerge(
  fcsFiles,
  comp = FALSE,
  transformMethod = c("autoLgcl", "cytofAsinh", "logicle", "arcsinh", "logAbs", "none"),
  scaleTo = NULL,
  mergeMethod = c("ceil", "all", "fixed", "min"),
  fixedNum = 2000,
  ...
)
}
\arguments{
\item{fcsFiles}{A vector of FCS file names.}

\item{comp}{If \verb{TRUE}, does compensation  by compensation matrix contained
in FCS. Agrument also accepts a compensation matrix to be applied.
Otherwise \verb{FALSE}.}

\item{transformMethod}{Data Transformation method, including \code{autoLgcl},
\code{cytofAsinh}, \code{logicle} and \code{arcsinh}, or \code{none} to
avoid transformation.}

\item{scaleTo}{Scale the expression to a specified range c(a, b), default is NULL.}

\item{mergeMethod}{Merge method for mutiple FCS expression data. cells can be
combined using one of the four different methods including \code{ceil},
\code{all}, \code{min}, \code{fixed}. The default option is \code{ceil},
up to a fixed number (specified by \code{fixedNum}) of cells are sampled
without replacement from each fcs file and combined for analysis.
\code{all}: all cells from each fcs file are combined for analysis.
\code{min}: The minimum number of cells among all the selected fcs files
are sampled from each fcs file and combined for analysis.
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled
(with replacement when the total number of cell is less than fixedNum)
from each fcs file and combined for analysis.}

\item{fixedNum}{The fixed number of cells to be extracted from each FCS file.}

\item{...}{Other arguments passed to \code{runExprsExtract}}
}
\value{
A matrix containing the merged expression data, with selected markers.
}
\description{
Apply preprocessing on each FCS file including compensation
  (for FCM data only) and transformation with selected markers, then expression
  matrix are extracted and merged using one of the methods, \code{all},
  \code{min}, \code{fixed} or \code{ceil}
}
\examples{


if (FALSE) {
  # See vignette tutorials for more information
  vignette("Quick_start", package = "flowSpy")

  # Path to your FCS files
  fcs.path <- "flowSpy-dataset/FCS/usecase2/"
  fcs.files <- paste0(fcs.path, "D", c(0,2,4,6,8,10), "-sub.fcs")

  # Merge FCS files, and each file contain 2000 cells
  set.seed(1)
  fcs.data <- runExprsMerge(fcs.files, comp = F, transformMethod = "none", fixedNum = 2000)
}


}
\references{
Hao Chen, Mai Chan Lau, Michael Thomas Wong, Evan W. Newell,
   Michael Poidinger, Jinmiao Chen. Cytofkit: A Bioconductor Package for
   an Integrated Mass Cytometry Data Analysis Pipeline. PLoS Comput Biol, 2016.
}
\seealso{
\code{\link{runExprsExtract}}
}
\author{
Chen Hao
}
