% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{runFastPCA}
\alias{runFastPCA}
\title{Calculate principal components in FSPY}
\usage{
runFastPCA(object, center = FALSE, scale. = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{center}{logical, a logical value indicating whether the variables
should be shifted to be zero centered. Alternately, a vector
of length equal the number of columns of x can be supplied.
The value is passed to scale. See \code{\link[gmodels]{fast.prcomp}}}

\item{scale.}{logical, a logical value indicating whether the
variables should be scaled to have unit variance before the
analysis takes place. The default is FALSE for consistency
with S, but in general scaling is advisable. Alternatively,
a vector of length equal the number of columns of x can be supplied.
The value is passed to scale. See \code{\link[gmodels]{fast.prcomp}}}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[gmodels]{fast.prcomp}} function}
}
\value{
An FSPY object with PCA
}
\description{
Calculate principal components in FSPY
}
\examples{

if (FALSE) {
fspy <- runFastPCA(fspy, verbose = TRUE)
}

}
\seealso{
\code{\link[gmodels]{fast.prcomp}}
}
