% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runKmeans}
\alias{runKmeans}
\title{runKmeans}
\usage{
runKmeans(
  object,
  k = 25,
  iter.max = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE,
  scale = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{k}{numeric. The number of clusters.}

\item{iter.max}{numeric. The maximum number of iterations allowed.}

\item{nstart}{numeric. If k is a number, how many random sets should be chosen.}

\item{algorithm}{character. Type of algorithm that will be choosen to calculate 
kmeans. Four algoritms are provided: Hartigan-Wong, Lloyd, Forgy, MacQueen.}

\item{trace}{logical or integer number.}

\item{scale}{logical. Whether to use scaled data in kmeans.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[stats]{kmeans}} function}
}
\value{
an FSPY object with kmeans.id in meta.data
}
\description{
Perform k-means clustering on a data matrix.
}
\examples{

if (FALSE) {
fspy <- runKmeans(fspy, k = 25, verbose = TRUE)
}

}
\seealso{
\code{\link[stats]{kmeans}}
}
