% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runSOM}
\alias{runSOM}
\title{calculation SOM in FSPY object}
\usage{
runSOM(
  object,
  xdim = 6,
  ydim = 6,
  rlen = 8,
  mst = 1,
  alpha = c(0.05, 0.01),
  radius = 1,
  init = FALSE,
  distf = 2,
  codes = NULL,
  importance = NULL,
  method = "euclidean",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{xdim}{Width of the grid.}

\item{ydim}{Hight of the grid.}

\item{rlen}{Number of times to loop over the training data for each MST}

\item{mst}{Number of times to build an MST}

\item{alpha}{Start and end learning rate}

\item{radius}{Start and end radius}

\item{init}{Initialize cluster centers in a non-random way}

\item{distf}{Distance function (1=manhattan, 2=euclidean, 3=chebyshev,
4=cosine)}

\item{codes}{Cluster centers to start with}

\item{importance}{array with numeric values. Parameters will be scaled
according to importance}

\item{method}{the distance measure to be used. This must be one of "euclidean",
"maximum", "manhattan", "canberra", "binary" or "minkowski".
Any unambiguous substring can be given. See \code{\link[stats]{dist}}}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[FlowSOM]{SOM}} function}
}
\value{
an FSPY object with som.id in FSPY object
}
\description{
Build a self-organizing map
}
\examples{
if (FALSE) {
fspy <- runSOM(fspy, xdim = 10, ydim = 10, verbose = TRUE)
}

}
\references{
This code is strongly based on the \code{\link[FlowSOM]{SOM}} function.
            Which is developed by Sofie Van Gassen, Britt Callebaut and Yvan Saeys (2018).
}
\seealso{
\code{\link{BuildSOM}}

\code{\link[FlowSOM]{SOM}}
}
