% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{runTSNE}
\alias{runTSNE}
\title{Calculate t-Distributed Stochastic Neighbor Embedding in FSPY}
\usage{
runTSNE(
  object,
  dims = 2,
  initial_dims = 50,
  perplexity = 30,
  theta = 0.5,
  check_duplicates = TRUE,
  pca = TRUE,
  max_iter = 1000,
  verbose = FALSE,
  is_distance = FALSE,
  Y_init = NULL,
  pca_center = TRUE,
  pca_scale = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{dims}{integer, Output dimensionality (default: 2)}

\item{initial_dims}{integer. the number of dimensions that should
be retained in the initial PCA step (default: 50). See \code{\link[Rtsne]{Rtsne}}}

\item{perplexity}{numeric. Perplexity parameter. See \code{\link[Rtsne]{Rtsne}}}

\item{theta}{numeric. Speed/accuracy trade-off (increase for less accuracy),
set to 0.0 for exact TSNE (default: 0.5). See \code{\link[Rtsne]{Rtsne}}}

\item{check_duplicates}{logical. Checks whether duplicates are present.
It is best to make sure there are no duplicates present and set this
option to FALSE, especially for large datasets (default: TRUE).
See \code{\link[Rtsne]{Rtsne}}}

\item{pca, max_iter, is_distance, Y_init, pca_center, pca_scale}{See \code{\link[Rtsne]{Rtsne}}}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[Rtsne]{Rtsne}} function}
}
\value{
An FSPY object
}
\description{
Calculate t-Distributed Stochastic Neighbor Embedding in FSPY
}
\examples{

if (FALSE) {

fspy <- runTSNE(fspy, dims = 2, verbose = TRUE)
fspy <- runTSNE(fspy, dims = 2, perplexity = 20, verbose = TRUE)

}

}
\references{
Maaten, L. Van Der, 2014. Accelerating t-SNE using Tree-Based
   Algorithms. Journal of Machine Learning Research, 15, p.3221-3245.

   van der Maaten, L.J.P. & Hinton, G.E., 2008. Visualizing High-Dimensional
   Data Using t-SNE. Journal of Machine Learning Research, 9, pp.2579-2605.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
