% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,CYT-method}
\alias{show,CYT-method}
\alias{show,}
\alias{show-method}
\title{CYT show method}
\usage{
\S4method{show}{CYT}(object)
}
\arguments{
\item{object}{An CYT object}
}
\value{
Cell number of CYT object
}
\description{
Prevents R from crashing by trying to print all slots of an CYT object  if a returned object is not stored in a variable.
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)
cyt


}
\keyword{internal}
