% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,FSPY-method}
\alias{show,FSPY-method}
\alias{show,}
\alias{show-method}
\title{FSPY show method}
\usage{
\S4method{show}{FSPY}(object)
}
\arguments{
\item{object}{An FSPY object}
}
\value{
Cell number of FSPY object
}
\description{
Prevents R from crashing by trying to print all slots of an FSPY object  if a returned object is not stored in a variable.
}
\examples{
if (FALSE) {
  fspy
}

}
\keyword{internal}
