% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateClustMeta}
\alias{updateClustMeta}
\title{Update clusters' meta information of CYT}
\usage{
updateClustMeta(object, verbose = TRUE)
}
\arguments{
\item{object}{An CYT object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An CYT object
}
\description{
Update clusters' meta information of CYT
}
\examples{

if (FALSE) {
cyt <- updateClustMeta(cyt)
}

}
