% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updatePlotMeta}
\alias{updatePlotMeta}
\title{Update plot meta information of CYT}
\usage{
updatePlotMeta(object, verbose = TRUE)
}
\arguments{
\item{object}{An CYT object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An CYT object
}
\description{
Update plot meta information of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

cyt <- updatePlotMeta(cyt)



}
