% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defLeafCells}
\alias{defLeafCells}
\title{definition of leaf cells}
\usage{
defLeafCells(object, leaf.cells = NULL, pseudotime.cutoff = 0, verbose = FALSE)
}
\arguments{
\item{object}{an FSPY object}

\item{leaf.cells}{character or numeric. Cell name of the root cells or
cluster.id of root.cells}

\item{pseudotime.cutoff}{numeric. Cutoff of pseudotime. Cells with pseudotime
over pseudotime.cutoff will be set to be leaf cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
definition of root cells
}
\examples{

if (FALSE) {
# Define leaf cells by cluster
fspy <- defLeafCells(fspy, leaf.cells = 1, verbose = TRUE)
fspy <- defLeafCells(fspy, leaf.cells = c(1,3), verbose = TRUE)

# Define root cells by cell names
cells <- test.meta.data$cell[which(test.meta.data$stage == "D10")]
cells <- as.character(cells)
fspy <- defLeafCells(fspy, leaf.cells = cells, verbose = TRUE)
}


}
