% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defRootCells}
\alias{defRootCells}
\title{definition of root cells}
\usage{
defRootCells(object, root.cells = NULL, verbose = FALSE)
}
\arguments{
\item{object}{a CYT object}

\item{root.cells}{vector. Cell name of the root cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object
}
\description{
definition of root cells
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

# Define root cells by cluster
cyt <- defRootCells(cyt, root.cells = 6, verbose = TRUE)
cyt <- defRootCells(cyt, root.cells = c(6,8), verbose = TRUE)

# Define root cells by cell names
meta.data <- fetchPlotMeta(cyt)
cells <- meta.data$cell[which(meta.data$stage == "D0")]
cells <- as.character(cells)
cyt <- defRootCells(cyt, root.cells = cells, verbose = TRUE)



}
