% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defRootCells}
\alias{defRootCells}
\title{definition of root cells}
\usage{
defRootCells(object, root.cells = NULL, verbose = FALSE)
}
\arguments{
\item{object}{an FSPY object}

\item{root.cells}{vector. Cell name of the root cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
definition of root cells
}
\examples{

if (FALSE) {
# Define root cells by cluster
fspy <- defRootCells(fspy, root.cells = 6, verbose = TRUE)
fspy <- defRootCells(fspy, root.cells = c(6,8), verbose = TRUE)

# Define root cells by cell names
cells <- test.meta.data$cell[which(test.meta.data$stage == "D0")]
cells <- as.character(cells)
fspy <- defRootCells(fspy, root.cells = cells, verbose = TRUE)
}


}
