% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchPlotMeta}
\alias{fetchPlotMeta}
\title{Fetching plot metadata of FSPY}
\usage{
fetchPlotMeta(object, markers = NULL, verbose = FALSE)
}
\arguments{
\item{object}{An FSPY object}

\item{markers}{vector. Makers fetched from expression matrix}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
a data.frame containing meta information for visualization
}
\description{
Fetching plot metadata of FSPY
}
\examples{

if (FALSE) {
plot.data <- fetchPlotMeta(fspy)
head(plot.data)

plot.data <- fetchPlotMeta(fspy, markers = c("CD43", "CD34"))
head(plot.data)
}



}
