% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotBranchHeatmap}
\alias{plotBranchHeatmap}
\title{Visualization heatmap of branch data of FSPY}
\usage{
plotBranchHeatmap(
  object,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  ...
)
}
\arguments{
\item{object}{An FSPY object}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of branch data of FSPY
}
\examples{

if (FALSE) {

plotBranchHeatmap(fspy)
plotBranchHeatmap(fspy, color = colorRampPalette(c("purple","white","yellow"))(100))
plotBranchHeatmap(fspy, cluster_row = FALSE)
plotBranchHeatmap(fspy, cluster_row = FALSE, cluster_col = FALSE)

}

}
