% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Visualization heatmap of data of CYT}
\usage{
plotHeatmap(
  object,
  markers = NULL,
  color = colorRampPalette(c("blue", "white", "red"))(100),
  scale = "row",
  downsize = 1000,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  ...
)
}
\arguments{
\item{object}{An CYT object}

\item{markers}{vector. markers to plot on the heatmap}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{downsize}{numeric. Cells size used to plot heatmap}

\item{cluster_rows}{logical. Whether rows should be clustered}

\item{cluster_cols}{logical. Whether columns should be clustered}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
ggplot2 figure
}
\description{
Visualization heatmap of data of CYT
}
\examples{

if (FALSE) {

plotHeatmap(cyt)
plotHeatmap(cyt, cluster_rows = T)
plotHeatmap(cyt, cluster_rows = T, clustering_method = "ward.D")
plotHeatmap(cyt, cluster_rows = T, cluster_cols = T)

}


}
