% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotPieCluster}
\alias{plotPieCluster}
\title{Visualization pie plot of cluster data of FSPY}
\usage{
plotPieCluster(
  object,
  item.use = c("PC_1", "PC_2"),
  cex.size = 1,
  size.by.cell.number = TRUE,
  main = "2D pie plot of FSPY",
  plot.theme = theme_bw()
)
}
\arguments{
\item{object}{An FSPY object}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{cex.size}{numeric. Size of the dot}

\item{size.by.cell.number}{logical. Whether to show size of cell number.}

\item{main}{character. Title of the plot.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\value{
ggplot2 figure
}
\description{
Visualization pie plot of cluster data of FSPY
}
\examples{

if (FALSE) {
# Runs only have more than two stages
plotPieCluster(fspy, cex.size = 0.5)

plotPieCluster(fspy, item.use = c("PC_1", "PC_2"), cex.size = 0.5)
plotPieCluster(fspy, item.use = c("PC_2", "PC_3"), cex.size = 0.5)

plotPieCluster(fspy, item.use = c("tSNE_1", "tSNE_2"), cex.size = 20)

plotPieCluster(fspy, item.use = c("DC_1", "DC_2"), cex.size = 0.5)

plotPieCluster(fspy, item.use = c("UMAP_1", "UMAP_2"), cex.size = 1)
plotPieCluster(fspy, item.use = c("UMAP_1", "UMAP_2"), cex.size = 1) +
   scale_fill_manual(values = c("#00599F","#FF3222","#009900",
                                "#FF9933","#FF99FF","#7A06A0"))
}

}
