% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{runUMAP}
\alias{runUMAP}
\title{Calculating UMAP}
\usage{
runUMAP(
  object,
  umap.config = umap.defaults,
  n_neighbors = 30,
  dims = 2,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{an FSPY object}

\item{umap.config}{object of class umap.config. See \code{\link[umap]{umap}}.}

\item{n_neighbors}{numeric. Number of neighbors}

\item{dims}{numeric. Dim of umap, you can also change it in umap.config.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Options to pass on to the \code{\link[umap]{umap}} function}
}
\value{
An FSPY object
}
\description{
Calculate Uniform Manifold Approximation and Projection in FSPY
}
\examples{

if (FALSE) {

fspy <- runUMAP(fspy, verbose = TRUE)
fspy <- runUMAP(fspy, n_neighbors = 20, verbose = TRUE)

}


}
\seealso{
\code{\link[umap]{umap}}
}
