% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{runWalk}
\alias{runWalk}
\title{Walk between root cells and leaf cells}
\usage{
runWalk(
  object,
  mode = c("undirected", "directed", "max", "min", "upper", "lower", "plus"),
  max.run.forward = 20,
  backward.walk = FALSE,
  max.run.backward = 20,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An FSPY object}

\item{mode}{character. Specifies how igraph should interpret the supplied matrix.
Possible values are: undirected, directed, upper, lower, max, min, plus. By 
default is undirected.}

\item{max.run.forward}{numeric. Maximum cycles of forward walk.}

\item{backward.walk}{logical. Whether to run backward walk.}

\item{max.run.backward}{numeric. Maximum cycles of backward walk.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to calculation function.}
}
\value{
An FSPY object
}
\description{
Walk between root cells and leaf cells
}
\examples{

if (FALSE) {
  fspy <- runWalk(fspy, verbose = TRUE)
  fspy <- runWalk(fspy, backward.walk = FALSE, verbose = TRUE)
}



}
