% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateClustMeta}
\alias{updateClustMeta}
\title{Update clusters' meta information of CYT}
\usage{
updateClustMeta(object, verbose = FALSE)
}
\arguments{
\item{object}{A CYT object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
A CYT object
}
\description{
Update clusters' meta information of CYT
}
\examples{

cyt.file <- system.file("extdata/cyt.rds", package = "CytoTree")
cyt <- readRDS(file = cyt.file)

cyt <- updateClustMeta(cyt)


}
