% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateClustMeta}
\alias{updateClustMeta}
\title{Update clusters' meta information of FSPY}
\usage{
updateClustMeta(object, verbose = TRUE)
}
\arguments{
\item{object}{An FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
Update clusters' meta information of FSPY
}
\examples{

if (FALSE) {
fspy <- updateClustMeta(fspy)
}

}
