% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updatePlotMeta}
\alias{updatePlotMeta}
\title{Update plot meta information of FSPY}
\usage{
updatePlotMeta(object, verbose = TRUE)
}
\arguments{
\item{object}{An FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
Update plot meta information of FSPY
}
\examples{

if (FALSE) {
fspy <- updatePlotMeta(fspy)
}


}
